+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -330,7 +330,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KTX87DJX8B2J6S5HF1SJVX6K"
+        "01KTXDE765Y02GGG1QSRNETJ59"
     )
     r"""The unique identifier of the data source"""