+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KTX87DJV779XXVR1V6HJYDP2"
+        "01KTXDE762JD0X71PATXD9YBTS"
     )
     r"""The unique identifier of the data source"""