+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -87,7 +87,7 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-12T06:27:06.026Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-12T07:58:13.213Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")