+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatapointop.py
@@ -1445,7 +1445,7 @@ class UpdateDatapointEvaluations3(BaseModel):
 
     source: Optional[UpdateDatapointEvaluationsDatasetsResponseSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-06-12T06:27:34.717Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-06-12T07:58:40.420Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1513,7 +1513,7 @@ class UpdateDatapointEvaluations2(BaseModel):
 
     source: Optional[UpdateDatapointEvaluationsDatasetsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-06-12T06:27:34.716Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-06-12T07:58:40.419Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1581,7 +1581,7 @@ class UpdateDatapointEvaluations1(BaseModel):
 
     source: Optional[UpdateDatapointEvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-06-12T06:27:34.716Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-06-12T07:58:40.419Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1684,7 +1684,7 @@ class UpdateDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-12T06:27:06.026Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-12T07:58:13.213Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")