+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KTX87DK03CA34TT0MHJJR7R1"
+        "01KTXDE767SRAEB1M48KX3NFQM"
     )
     r"""The unique identifier of the data source"""