+++ packages/orq-rc/RELEASES.md
@@ -8659,3 +8659,13 @@ Based on:
 - [python v4.11.0-rc.15] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.15] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.15 - packages/orq-rc
+
+## 2026-06-12 08:00:45
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.777.2 (2.903.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.16] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.16] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.16 - packages/orq-rc